+++ lib/graphql/persisted_queries/compiled_queries/query_patch.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+module GraphQL
+  module PersistedQueries
+    module CompiledQueries
+      # Patches GraphQL::Query to support compiled queries
+      module QueryPatch
+        def persisted_query_not_found?
+          @persisted_query_not_found
+        end
+
+        def prepare_ast
+          return super unless @context[:extensions]
+
+          @document = resolver.fetch
+          not_loaded_document = @document.nil?
+
+          @persisted_query_not_found = not_loaded_document && query_string.nil?
+
+          super.tap do
+            resolver.persist(query_string, @document) if not_loaded_document && query_string
+          end
+        end
+
+        private
+
+        def resolver
+          @resolver ||= Resolver.new(@schema, @context[:extensions])
+        end
+      end
+    end
+  end
+end