+++ lib/graphql/persisted_queries/store_adapters/base_store_adapter.rb
@@ -12,15 +12,18 @@ module GraphQL
           @name = :base
         end
 
-        def fetch_query(hash)
-          fetch(hash).tap do |result|
+        def fetch_query(hash, compiled_query: false)
+          key = build_key(hash, compiled_query)
+
+          fetch(key).tap do |result|
             event = result ? "cache_hit" : "cache_miss"
             trace("fetch_query.#{event}", adapter: @name)
           end
         end
 
-        def save_query(hash, query)
-          trace("save_query", adapter: @name) { save(hash, query) }
+        def save_query(hash, query, compiled_query: false)
+          key = build_key(hash, compiled_query)
+          trace("save_query", adapter: @name) { save(key, query) }
         end
 
         protected
@@ -40,6 +43,13 @@ module GraphQL
           elsif block_given?
             yield
           end
+        end
+
+        private
+
+        def build_key(hash, compiled_query)
+          key = "#{RUBY_ENGINE}-#{RUBY_VERSION}:#{GraphQL::VERSION}:#{hash}"
+          compiled_query ? "compiled:#{key}" : key
         end
       end
     end